Use float64 rather than float32

+++ src/modules/proto_dhcpv4/dpc_config.h
@@ -22,7 +22,7 @@ struct dpc_config_s {
 	bool debug_basename;             //!< Print only file base name.
 	bool log_timestamp;              //!< Add timestamp to log messages.
 
-	float progress_interval;         //<! Time interval between periodic progress statistics.
+	double progress_interval;        //<! Time interval between periodic progress statistics.
 	bool pr_stat_timestamp;          //!< Add timestamp to progress statistics.
 	bool pr_stat_per_input;          //<! Print per-input progress statistics (if multiple input).
 	uint32_t pr_stat_per_input_max;  //<! Max number of input items shown in progress statistics.
@@ -38,11 +38,11 @@ struct dpc_config_s {
 	int32_t packet_trace_lvl;        //<! Packet trace level (0: none, 1: header, 2: and attributes, 3: and encoded hex data).
 	bool packet_trace_elapsed;       //<! Prefix packet trace with elapsed time.
 	bool packet_trace_timestamp;     //<! Prefix packet trace with current timestamp.
-	float request_timeout;           //<! Max time waiting for a reply to a request we've sent.
+	double request_timeout;          //<! Max time waiting for a reply to a request we've sent.
 	uint32_t retransmit_max;         //<! Max retransmissions of a request not replied to (not including first packet).
 
-	float rate_limit;                //<! Limit rate/s of sessions initialized from input (all transactions combined).
-	float duration_start_max;        //<! Limit duration for starting new input sessions.
+	double rate_limit;               //<! Limit rate/s of sessions initialized from input (all transactions combined).
+	double duration_start_max;       //<! Limit duration for starting new input sessions.
 	uint32_t input_num_use;          //<! Max number of uses of each input item (default: unlimited in template mode, 1 otherwise).
 	uint32_t session_max_num;        //<! Limit number of sessions initialized from input items.
 	uint32_t session_max_active;     //<! Max number of session packets sent concurrently (default: 1).
